% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencefunctions.R
\name{completeFun}
\alias{completeFun}
\title{Drop NAs by Columns}
\usage{
completeFun(data, desiredCols)
}
\description{
Remove NAs based on specified columns in the data
 @param data data.frame object of observations
 @param desiredCols list of columns from which incomplete cases should be dropped
 
 @return dataframe with removed observations
 @examples 
 data<-data.frame(a=1:4,b=c("a", "b", "c", "d"), c=c(NA, "keep", NA, "keep))
 completeFun(data,c("c"))
 
 @export
}
